update web cms

+++ database/migrations/2021_08_10_075323_table_user.php
@@ -18,7 +18,7 @@ class TableUser extends Migration
             $table->string("prestasi")->after("email")->nullable();
             $table->string("posisi_terakhir_bekerja")->after("email")->nullable();
             $table->string("tempat_terakhir_bekerja")->after("email")->nullable();
-            $table->double("lama_terakhir_bekerja")->after("email")->nullable();
+            $table->integer("lama_terakhir_bekerja")->after("email")->nullable();
         });
     }
