+++ database/migrations/2021_08_10_161034_table_pekerjaan.php
@@ -14,7 +14,7 @@ class TablePekerjaan extends Migration
             $table->string("nama_pekerjaan");
             $table->longText("deskripsi");
             $table->string("nama_perusahaan")->nullable();
-            $table->double("gaji")->nullable();
+            $table->integer("gaji")->nullable();
             $table->string("logo_perusahaan_path", 2048)->nullable();
             $table->string("foto_lowongan", 2048)->nullable();
             $table->string("tentang_pembuka_lowongan")->nullable();
